telegram_bot/requirements-bot.txt
# Phase 2 — Telegram Bot Dependencies
# Separate from FastAPI backend requirements

python-telegram-bot==21.6
httpx==0.27.0
redis==5.0.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
apscheduler==3.10.4
pydantic==2.7.4
python-dotenv==1.0.1
structlog==24.4.0
